[ETCM-921] Introduce TypedTransaction ADT
